requirements.txt

pyTelegramBotAPI
flask
gunicorn
python-binance
